Don't deploy contracts during the build process (done manually)
The contracts shouldn't be deployed to QA or prod during the automatic build process.
This is because it costs money, it's quite rare that we change the
contracts and also the contract address files (shared/src/ethlance/shared/smart_contracts_{qa,prod}.cljs)
need to be committed to version control to maintain the addresses.

There are ways for doing it via CI but as of now it's not worth the
effort