Register individual handler fn-s in syncer
Previously the syncer registered handler fn that could handle all the
events and registered it for all the events and the selection was made
inside that function.

This is duplication and potentially causes errors. This approach is
narrower - single handler for event gets registered (after wrapping it
to give it DB connection and calling it inside try-catch). Similar
approach is used in StreamTide